+++ backend/deleteText.php
@@ -1,14 +1,30 @@
 <?php
+//deleting a text
 
+//check if logged in
+session_start();
+
+if (!isset($_SESSION['user_id'])){
+    $data = array(
+        'error'=> 'You are not allowed here!'
+    );
+    header('Location: ../index.php');
+    die();
+
+}
+
+//check if there is a text id
 if (!isset($_GET['id'])){
     header('Location: ../about.php');
 }
 
-
+//prepare variable
 $textid = $_GET['id'];
 
+//connection to database
 include_once 'pdo-connect.php';
 
+//delete text data from database
 try{
     $stmt = $conn->prepare("DELETE FROM texts WHERE id = :textid ");
     $stmt->bindParam(':textid', $textid);
